HW/HW_09/pyproject.toml
[tool.poetry]
name = "hw-08"
version = "0.1.0"
description = ""
authors = ["AndriiRohovyk <77483613+AndreyRogovik@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pymongo = {extras = ["srv"], version = "^4.6.0"}
mongoengine = "^0.27.0"
python-dateutil = "^2.8.2"
pika = "^1.3.2"
faker = "^20.1.0"
lxml = "^4.9.3"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
